move sanitizer.ts to /util folder
